nbp client, error handling, few cleanups and fixes

+++ application/src/main/java/net/hubek/nn/controller/response/AccountResponse.java
@@ -2,6 +2,13 @@ package net.hubek.nn.controller.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
-public record AccountResponse(String id, String firstName, String lastName, long plnBalance, long usdBalance) {
+public record AccountResponse(
+        String id,
+        String firstName,
+        String lastName,
+        BigDecimal plnBalance,
+        BigDecimal usdBalance) {
 }
